Renamed tasks to tasks-handlers. Build error fix.
