+++ example/src/index.html
@@ -1,3 +1,3 @@
 <html>
-HA > 
-</html>
+HA >
+</html>
