--- src/watchers/watcher-contracts.ts
@@ -1,7 +1,0 @@
-import { TaskFunction, Globs } from 'gulp/contracts';
-
-export interface WatchTask {
-    Name: string;
-    TaskFunction: TaskFunction;
-    Globs: Globs;
-}
